Python/rons-bot/requirements.txt
discord==1.0.1
discord.py==1.6.0
google-api-core==1.25.0
google-auth==1.24.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
googletrans==4.0.0rc1
python-dotenv==0.15.0
